Add animation loading and keyframe blending for v2

+++ src/rendering/v2/game_counter.h
@@ -4,7 +4,7 @@
 namespace hyperion::v2 {
 
 struct GameCounter {
-    using TickUnit = double;
+    using TickUnit = float;
 
     TickUnit last_tick{};
     TickUnit tick{};
